checkpoints/finetune_dna_human/code/01_11-15:32:28/requirements.txt
numpy==1.26.4
torch
torchvision
roma
gradio
matplotlib
tqdm
opencv-python
scipy
einops
trimesh
tensorboard
pyglet<2
huggingface-hub[torch]>=0.22
viser
gradio
lpips
hydra-core
pillow==10.3.0
h5py
accelerate
transformers
scikit-learn